requirements.txt
absl-py==1.0.0
albumentations==1.1.0
altair==4.1.0
anyio==3.3.4
argcomplete==1.12.3
argon2-cffi==21.1.0
astor==0.8.1
attrs==21.2.0
Babel==2.9.1
backcall==0.2.0
backports.entry-points-selectable==1.1.1
backports.zoneinfo==0.2.1
base58==2.1.1
bleach==4.1.0
blinker==1.4
CacheControl==0.12.10
cachetools==4.2.4
cachy==0.3.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
cleo==0.8.1
click==7.1.2
clikit==0.6.2
crashtest==0.3.1
cryptography==36.0.0
cycler==0.11.0
Cython==0.29.26
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
distlib==0.3.4
efficientnet-pytorch==0.6.3
entrypoints==0.3
filelock==3.4.0
future==0.18.2
gitdb==4.0.9
GitPython==3.1.24
google-auth==1.35.0
google-auth-oauthlib==0.4.6
grad-cam==1.3.5
grpcio==1.42.0
html5lib==1.1
idna==3.3
imageio==2.13.3
importlib-metadata==1.7.0
importlib-resources==5.4.0
ipykernel==6.5.0
ipython==7.29.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.0
jeepney==0.7.1
Jinja2==3.0.2
joblib==1.1.0
json5==0.9.6
jsonschema==4.2.1
jupyter-client==6.1.12
jupyter-core==4.9.1
jupyter-server==1.11.2
jupyterlab==3.2.2
jupyterlab-pygments==0.1.2
jupyterlab-server==2.8.2
jupyterlab-widgets==1.0.2
keyring==21.8.0
kiwisolver==1.3.2
lockfile==0.12.2
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.2.1
matplotlib-inline==0.1.3
mistune==0.8.4
msgpack==1.0.3
munch==2.5.0
nbclassic==0.3.4
nbclient==0.5.4
nbconvert==6.2.0
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.6.3
notebook==6.4.5
numpy==1.21.4
oauthlib==3.1.1
opencv-python==4.5.1.48
opencv-python-headless==4.5.4.60
packaging==20.9
pandas==1.1.5
pandocfilters==1.5.0
parso==0.8.2
pastel==0.2.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
pkginfo==1.8.2
platformdirs==2.4.0
poetry==1.1.12
poetry-core==1.0.7
pool==0.1.2.dev0
pretrainedmodels==0.7.4
prometheus-client==0.12.0
prompt-toolkit==3.0.22
protobuf==3.19.1
ptyprocess==0.7.0
pyarrow==6.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydeck==0.7.1
pydensecrf @ git+https://github.com/lucasb-eyer/pydensecrf.git@0d53acbcf5123d4c88040fe68fbb9805fc5b2fb9
Pygments==2.10.0
pylev==1.4.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.2.0
PyYAML==6.0
pyzmq==22.3.0
qudida==0.0.4
requests==2.26.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rsa==4.8
scikit-image==0.19.0
scikit-learn==0.24.2
scipy==1.7.3
SecretStorage==3.3.1
segmentation-models-pytorch==0.2.1
Send2Trash==1.8.0
shellingham==1.4.0
six==1.16.0
smmap==5.0.0
sniffio==1.2.0
streamlit==1.1.0
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
terminado==0.12.1
testpath==0.5.0
threadpoolctl==3.0.0
tifffile==2021.11.2
timm==0.4.12
toml==0.10.2
tomlkit==0.7.2
toolz==0.11.2
torch==1.10.1
torchvision==0.7.0
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
ttach==0.0.3
typing-extensions==3.10.0.2
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.7
validators==0.18.2
virtualenv==20.10.0
voila==0.2.16
watchdog==2.1.6
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.1
Werkzeug==2.0.2
widgetsnbextension==3.5.2
zipp==3.6.0
